requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
attrs==23.1.0
autopep8==2.0.2
beautifulsoup4==4.12.2
black==23.3.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
dateutils==0.6.12
discord.py==2.2.3
et-xmlfile==1.1.0
frozenlist==1.3.3
h11==0.14.0
httpcore==0.17.0
httpx==0.24.0
idna==3.4
multidict==6.0.4
mypy-extensions==1.0.0
openpyxl==3.1.2
packaging==23.1
pathspec==0.11.1
platformdirs==3.5.0
pycodestyle==2.10.0
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
tomli==2.0.1
tqdm==4.65.0
tweety-ns @ git+https://github.com/mahrtayyab/tweety@a79444b473770c944f8cc7b053bf06e972552e3b
yarl==1.9.2
